+++ packages/expo-dev-launcher/ios/SwiftUI/Navigation/Navigation.swift
@@ -58,9 +58,7 @@ struct DevLauncherNavigationHeader: View {
         } else {
           ZStack {
             Circle()
-            #if !os(tvOS)
-              .fill(Color(.systemGray6))
-            #endif
+              .fill(Color.expoSystemColors.systemGray6.color)
               .frame(width: 36, height: 36)
 
             Image("user-icon", bundle: getDevLauncherBundle())
@@ -72,9 +70,7 @@ struct DevLauncherNavigationHeader: View {
     }
     .padding(.horizontal)
     .padding(.vertical, 8)
-    #if !os(tvOS)
-    .background(Color(.systemBackground))
-    #endif
+    .background(Color.expoSystemColors.systemBackground.color)
   }
 
   @ViewBuilder
@@ -103,9 +99,7 @@ struct DevLauncherNavigationHeader: View {
           .aspectRatio(contentMode: .fill)
       } placeholder: {
         Circle()
-        #if !os(tvOS)
-          .fill(Color(.systemGray5))
-        #endif
+          .fill(Color.expoSystemColors.systemGray5.color)
           .overlay(
             Image(systemName: "person")
               .font(.system(size: 16))